+++ src/pages/external/resetPassword/component/index.tsx
@@ -1,0 +1,1 @@
+export * from './resetPasswordForm'
